Port sale_start_end_dates and sale_rental (#353)
