+++ sale_start_end_dates/demo/sale_demo.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-    Copyright (C) 2014-2015 Akretion (http://www.akretion.com/)
-    @author: Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
+  © 2014-2016 Akretion (http://www.akretion.com/)
+  @author: Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 
-<openerp>
-<data noupdate="1">
+<odoo noupdate="1">
 
 
 <!-- Add admin and demo user to group "Properties on lines", so that they have
@@ -20,5 +19,4 @@ access to the form view of sale order lines -->
 </record>
 
 
-</data>
-</openerp>
+</odoo>
