+++ sale_start_end_dates/views/sale_order.xml
@@ -1,13 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <!--
-    Copyright (C) 2014-2015 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
+  © 2014-2016 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 
-<openerp>
-<data>
+<odoo>
 
 
 <record id="view_order_form" model="ir.ui.view">
@@ -34,21 +32,8 @@
             <field name="default_start_date"/>
             <field name="default_end_date"/>
         </field>
-        <xpath expr="//field[@name='order_line']/form//field[@name='product_id']" position="attributes">
-            <attribute name="context_default_start_date">{'default_start_date': parent.default_start_date}</attribute>
-            <attribute name="context_default_end_date">{'default_end_date': parent.default_end_date}</attribute>
-        </xpath>
-        <xpath expr="//field[@name='order_line']/tree/field[@name='product_id']" position="attributes">
-            <attribute name="context_default_start_date">{'default_start_date': parent.default_start_date}</attribute>
-            <attribute name="context_default_end_date">{'default_end_date': parent.default_end_date}</attribute>
-        </xpath>
-        <xpath expr="//field[@name='order_line']/form//field[@name='product_uom_qty']" position="attributes">
-            <attribute name="context_default_start_date">{'default_start_date': parent.default_start_date}</attribute>
-            <attribute name="context_default_end_date">{'default_end_date': parent.default_end_date}</attribute>
-        </xpath>
     </field>
 </record>
 
 
-</data>
-</openerp>
+</odoo>
